채팅 업데이트

+++ back/src/models/club-posts/club-posts.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { In, Repository } from 'typeorm';
 import { CreatePostBody } from './dto/create-post.dto';
 import { UpdatePostBody } from './dto/update-post.dto';
 import { ClubPosts } from './entities/club-posts.entity';
@@ -20,15 +20,17 @@ export class ClubPostsService {
     const clubPost = new ClubPosts();
     clubPost.title = body.title;
     clubPost.content = body.content;
+    clubPost.showStatus = body.showStatus;
     clubPost.ClubId = clubId;
     clubPost.ClubMemberId = clubMemberId;
     await this.clubPostsRepository.save(clubPost);
     return;
   }
 
-  async findAllPosts(clubId: number): Promise<ClubPosts[]> {
+  async findAllPosts(clubId: number, showStatus: string): Promise<ClubPosts[]> {
+    const showStatusArr = showStatus.split(',');
     const result = await this.clubPostsRepository.find({
-      where: { ClubId: clubId },
+      where: { ClubId: clubId, showStatus: In(showStatusArr) },
       relations: ['ClubMember'],
       order: {
         createdAt: 'DESC',
@@ -51,6 +53,10 @@ export class ClubPostsService {
     });
     clubPost.title = body.title;
     clubPost.content = body.content;
+    clubPost.showStatus = body.showStatus;
+    if (body.createdAt) clubPost.createdAt = body.createdAt;
+    clubPost.updatedAt = body.updatedAt;
+
     await this.clubPostsRepository.save(clubPost);
     return;
   }
@@ -59,6 +65,9 @@ export class ClubPostsService {
     const clubPost = await this.clubPostsRepository.findOne({
       where: { id: postId },
     });
+    if (!clubPost) {
+      throw new ConflictException('db에 존재하지 않습니다.');
+    }
     await this.clubPostsRepository.remove(clubPost);
     return;
   }
